requirements.txt
# Telegram client library
pyrotgfork

# For tgcrypto speedups
tgcrypto

# PostgreSQL adapter for Python
psycopg2-binary

# For reading .env files
python-dotenv
